Create README.md
There was a problem - with long podcasts on YouTube, the video itself is not needed and I want to turn off the screen. 
The bot solves the problem with the inability to lock the screen when listening to a video on YouTube. 
This bot does the following: it takes a link to a YouTube video and cuts the audio track.
I also covered the code with unit tests for the correct operation of all functions.